+++ tests/Integration/Pricing/PricingTest.php
@@ -6,7 +6,7 @@
  * Time: 18:31
  */
 
-namespace Tests\Unit;
+namespace Tests\Integration;
 
 use LKDev\HetznerCloud\Models\Datacenters\Datacenters;
 use LKDev\HetznerCloud\Models\Prices\Prices;