+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (XX.XX.XXXX)
  + Add `automount` and `format` Parameters to `LKDev\HetznerCloud\Models\Volumes\Volumes` - `create`
  + Add `created` property to `LKDev\HetznerCloud\Models\FloatingIps\FloatingIp`
+ + Improve test coverage
 
 ## 1.6.0 (08.05.2019)
 + Setting the UserAgent is now possible with `LKDev\HetznerCloud\HetznerAPIClient` - `setUserAgent()`